Bounding volumes refactoring
