+++ include/geSG/BoundingVolume.h
@@ -1,0 +1,27 @@
+#pragma once
+//#include <geSG/Export.h>
+
+#include <geUtil/Ray.h>
+
+#include <glm/glm.hpp>
+#include <memory>
+#include <algorithm>
+
+namespace ge
+{
+   namespace sg
+   {
+      /**
+       * Base class for general bounding volume.
+       * Class is currently empty.
+       */   
+
+      class BoundingVolume
+      {
+      public:
+         //virtual void expand(const BoundingVolume* BV) = 0;
+      };
+
+      
+   }//namesace sg
+}//namespace ge
