+++ include/geSG/RayTriangleIntersector.h
@@ -20,7 +20,7 @@ namespace ge{
          static bool intersects(const util::Ray & ray, ge::sg::Triangle& triag);
          static float computeIntersection(const util::Ray & ray, ge::sg::Triangle& triag);
 
-         virtual bool intersects() override;
+         virtual bool intersects() const override;
 
          util::Ray ray;
          glm::vec3 v0, v1, v2;
